requirements.txt
evdev
RPi.GPIO